Implemented threading for f-score and a new data handler
Added a new file, splitter.py, to handle the data cleanup separately from the actual f-score file. It has
functionality for data collection rate, several split locations, and margin size.

threaded_f_score.py has been streamlined thanks to it no longer handling the data itself (though this also means
it is now dependent on splitter.py or manual data cleaning). Threads have been implemented along with full use
of the map-reduce method. A redundancy check has been included to ensure that the statistics calculations for a
block of data only need to be run once despite each block showing up several times in the pair list.

Lastly a note for future changes: the parallel processing implementation used seems to work best with a smaller
number of threads (~15) than originally intended, so the implementation should be updated to allow for a much more
intensive threading (the goal was 64 or 128 threads).